Online/offline switching, bugfixes for saving categories and notes

+++ classes/note.php
@@ -16,14 +16,10 @@ class Note {
 		"expired"
 	];
 
-	private $base_path;
-
 
 	// create note from data array
 	// [title, text, created, modified, start, expired]
 	public function __construct($data = []) {
-
-		$this->base_path = Config::config("path_content") . 'news/';
 		$this->init($data);
 	}
 
@@ -59,7 +55,6 @@ class Note {
 	// save note to file
 	public function save($category, $file = false) {
 
-
 		if (!$this->data["created"]) {
 			$this->data["created"] = time();
 		}
@@ -75,11 +70,11 @@ class Note {
 			$file = uniqid();
 		}
 
-		$cat_path = $this->base_path . $category . '/';
-		$old_cat_path = $this->base_path . Session::param("news_old_cat") . '/';
+		$cat_path = Category::base($category);
+		$old_cat_path = Category::base(Session::param("news_old_cat"));
 
 		// add category if don't exist
-		$this->add_category($category);
+		Category::add_category($category);
 
 
 		// has file and category
@@ -89,13 +84,14 @@ class Note {
 			$ini = new Ini();
 			$ini->load($this->data);
 
+
 			// save
 			if ($ini->save($cat_path . $file . ".ini")) {
 
 				Message::success("note_saved");
 
 				// if saved and moved to other category -> clear old entry
-				if ($cat_path != $old_cat_path) {
+				if ($old_cat_path && $cat_path != $old_cat_path) {
 					Session::set_param("news_cat", $category);
 					unlink($old_cat_path . $file . ".ini");
 				}
@@ -105,17 +101,6 @@ class Note {
 			else {
 				Message::failure("note_save_failure");
 			}
-		}
-	}
-
-
-	// add category if don't exists
-	public function add_category($category) {
-
-		$path = $this->base_path . $category . "/";
-
-		if (!file_exists($path)) {
-			mkdir($path);
 		}
 	}
